Added remaining mathematical Operators, Needs testing

+++ src/main/java/com/csci6461/MathMR.java
@@ -29,10 +29,10 @@ public class MathMR extends Instruction{
         /* Save trap code to args array */
         int[] args = new int[4];
 
-        args[0] = Integer.parseInt(gpr,2);
-        args[1] = Integer.parseInt(ixr,2);
-        args[2] = Integer.parseInt(i,2);
-        args[3] = Integer.parseInt(address,2);
+        args[0] = (short)Integer.parseInt(gpr,2);
+        args[1] = (short)Integer.parseInt(ixr,2);
+        args[2] = (short)Integer.parseInt(i,2);
+        args[3] = (short)Integer.parseInt(address,2);
 
         /* Return args */
         return args;
